+++ NEWS.md
@@ -1,3 +1,11 @@
+# pilot2wrappers 0.10.0
+
+- Add usage guide for application (#61)
+- Remove treatment variables from the `adtte` set prior to generating filters in KM plot module (#60)
+- Ensure data filters within the KM-Plot module are only applied to data sets inside that module, and not on the remaining modules (#59)
+* Add specific notes regarding filter usage and overlapping variables inside the KM module per user feedback
+* Fix typos in ADRG Quarto document
+
 # pilot2wrappers 0.9.0
 
 - Fix RMSE calculation in efficacy table to match result in Pilot 1